feat: 카테고리 제거 권한을 생성자가 아닌 ADMIN 역할의 유저가 할 수 있도록 변경

+++ backend/src/main/java/com/allog/dallog/domain/category/application/CategoryService.java
@@ -121,10 +121,12 @@ public class CategoryService {
 
     @Transactional
     public void delete(final Long memberId, final Long id) {
-        categoryRepository.validateExistsByIdAndMemberId(id, memberId);
         Category category = categoryRepository.getById(id);
 
         validateNotPersonalCategory(category);
+
+        CategoryRole role = categoryRoleRepository.getByMemberIdAndCategoryId(memberId, category.getId());
+        role.validateAuthority(CategoryAuthority.DELETE_CATEGORY);
 
         scheduleRepository.deleteByCategoryIdIn(List.of(id));
         subscriptionRepository.deleteByCategoryIdIn(List.of(id));
